+++ tests/Unit/Expressions/Operators/LessThanOrEqualTest.php
@@ -19,17 +19,18 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace WikibaseSolutions\CypherDSL\Tests\Unit\Expressions;
+namespace WikibaseSolutions\CypherDSL\Tests\Unit\Expressions\Operators;
 
 use PHPUnit\Framework\TestCase;
 use TypeError;
+use WikibaseSolutions\CypherDSL\Tests\Unit\Expressions\TestHelper;
 use WikibaseSolutions\CypherDSL\Expressions\GreaterThanOrEqual;
 use WikibaseSolutions\CypherDSL\Expressions\LessThanOrEqual;
 use WikibaseSolutions\CypherDSL\Types\AnyType;
 use WikibaseSolutions\CypherDSL\Types\PropertyTypes\NumeralType;
 
 /**
- * @covers \WikibaseSolutions\CypherDSL\Expressions\LessThanOrEqual
+ * @covers \WikibaseSolutions\CypherDSL\Expressions\Operators\LessThanOrEqual
  */
 class LessThanOrEqualTest extends TestCase
 {
